requirements.txt
mkdocs-material
mkdocs-git-revision-date-localized-plugin
mkdocs-awesome-pages-plugin
pyyaml
